fix(api): enhance deployment deletion logic to remove associated workflow URLs and versions

+++ server/api/internal/usecase/interactor/deployment.go
@@ -268,14 +268,33 @@ func (i *Deployment) Delete(ctx context.Context, deploymentID id.DeploymentID) (
 		return err
 	}
 
-	if url, _ := url.Parse(dep.WorkflowURL()); url != nil {
-		if err := i.file.RemoveWorkflow(ctx, url); err != nil {
+	if dep.Project() != nil {
+		versions, err := i.deploymentRepo.FindVersions(ctx, dep.Workspace(), dep.Project())
+		if err != nil {
 			return err
 		}
-	}
 
-	if err := i.deploymentRepo.Remove(ctx, deploymentID); err != nil {
-		return err
+		for _, version := range versions {
+			if url, _ := url.Parse(version.WorkflowURL()); url != nil {
+				if err := i.file.RemoveWorkflow(ctx, url); err != nil {
+					return err
+				}
+			}
+
+			if err := i.deploymentRepo.Remove(ctx, version.ID()); err != nil {
+				return err
+			}
+		}
+	} else {
+		if url, _ := url.Parse(dep.WorkflowURL()); url != nil {
+			if err := i.file.RemoveWorkflow(ctx, url); err != nil {
+				return err
+			}
+		}
+
+		if err := i.deploymentRepo.Remove(ctx, deploymentID); err != nil {
+			return err
+		}
 	}
 
 	tx.Commit()
